move feature helper methods into separate file
